Refactor: Extract useIsInvalidLink hook to shared module
Moves the `useIsInvalidLink` hook to a shared location to reduce code duplication between navigation-link and navigation-submenu blocks.

## Changes

- Created `navigation-link/shared/use-is-invalid-link.js` with shared hook
- Updated `navigation-link/shared/index.js` to export the new hook
- Refactored `navigation-link/edit.js` to import from shared
- Refactored `navigation-submenu/edit.js` to import from shared
- Removed duplicate hook definitions (65+ lines of duplicated code)
- Cleaned up unused imports (coreStore, useBlockEditingMode)

This ensures both blocks use identical validation logic and makes future maintenance easier.
