+++ packages/block-library/src/navigation-link/shared/index.js
@@ -13,3 +13,4 @@ export {
 } from './use-entity-binding';
 export { LinkUI } from '../link-ui';
 export { useHandleLinkChange } from './use-handle-link-change';
+export { useIsInvalidLink } from './use-is-invalid-link';
